Removing unprocessed content plugin code from view
When we don't need content plugin to be processed in form view (in intro, outro etc) 
and this option is choosen in form admin, then we shouldn't see anything but we see the code (e.g {fabrik view=list id=# table___element=[rowid]} ) that has no sense. With this - after regular {placeholders} are parsed, everything that is between {} and not parsed is just removed from view.  