Re-making pk field as internalid by default
During the enhancement that allowed to determine the default element types for different db field types we forgot to pay attention to primary keys that MUST be internalid by default anyway. So when new list were created on db table and default types determined by user then also pk fields were created as 'field' elements and needed to be changed manually. This is fixed now.